+++ cmd/internal/cli/build.go
@@ -1,5 +1,5 @@
 // Copyright (c) 2020, Control Command Inc. All rights reserved.
-// Copyright (c) 2018-2022, Sylabs Inc. All rights reserved.
+// Copyright (c) 2018-2023, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -41,6 +41,7 @@ var buildArgs struct {
 	isJSON        bool
 	noCleanUp     bool
 	noTest        bool
+	noSetgroups   bool
 	remote        bool
 	sandbox       bool
 	update        bool
@@ -186,6 +187,16 @@ var buildFakerootFlag = cmdline.Flag{
 	EnvKeys:      []string{"FAKEROOT"},
 }
 
+// --no-setgroups
+var buildNoSetgroupsFlag = cmdline.Flag{
+	ID:           "buildNoSetgroupsFlag",
+	Value:        &buildArgs.noSetgroups,
+	DefaultValue: false,
+	Name:         "no-setgroups",
+	Usage:        "disable setgroups when entering --fakeroot user namespace",
+	EnvKeys:      []string{"NO_SETGROUPS"},
+}
+
 // -e|--encrypt
 var buildEncryptFlag = cmdline.Flag{
 	ID:           "buildEncryptFlag",
@@ -286,6 +297,7 @@ func init() {
 		cmdManager.RegisterFlagForCmd(&buildDisableCacheFlag, buildCmd)
 		cmdManager.RegisterFlagForCmd(&buildEncryptFlag, buildCmd)
 		cmdManager.RegisterFlagForCmd(&buildFakerootFlag, buildCmd)
+		cmdManager.RegisterFlagForCmd(&buildNoSetgroupsFlag, buildCmd)
 		cmdManager.RegisterFlagForCmd(&buildFixPermsFlag, buildCmd)
 		cmdManager.RegisterFlagForCmd(&buildJSONFlag, buildCmd)
 		cmdManager.RegisterFlagForCmd(&buildLibraryFlag, buildCmd)
@@ -331,6 +343,10 @@ var buildCmd = &cobra.Command{
 }
 
 func preRun(cmd *cobra.Command, args []string) {
+	if buildArgs.noSetgroups && !buildArgs.fakeroot {
+		sylog.Warningf("--no-setgroups only applies to --fakeroot builds")
+	}
+
 	if buildArgs.fakeroot && !buildArgs.remote {
 		fakerootExec(args)
 	}