+++ internal/pkg/runtime/launcher/native/launcher_linux.go
@@ -260,6 +260,14 @@ func (l *Launcher) Exec(ctx context.Context, image string, process string, args 
 	if l.cfg.Fakeroot {
 		l.cfg.Namespaces.User = true
 	}
+	// Allow optional skipping of setgroups in --fakeroot mode.
+	if l.cfg.NoSetgroups {
+		if l.cfg.Fakeroot {
+			l.engineConfig.SetNoSetgroups(l.cfg.NoSetgroups)
+		} else {
+			sylog.Warningf("--no-setgroups only applies to --fakeroot mode")
+		}
+	}
 
 	l.setCgroups(instanceName)